+++ src/Foundation/resources/database/migrations/2023_02_22_191444_add_shipping_method_and_customer_to_the_orders_table.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class () extends Migration {
+    public function up()
+    {
+        Schema::table('orders', function (Blueprint $table) {
+            $table->unsignedBigInteger('shipping_method_id')->nullable()->after('shipping_address_id');
+            $table->intOrBigIntBasedOnRelated('customer_id', Schema::connection(null), 'customers.id')->unsigned()->nullable()->after('user_id');
+
+            $table->foreign('shipping_method_id')->references('id')->on('shipping_methods');
+            $table->foreign('customer_id')->references('id')->on('customers');
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('orders', function (Blueprint $table) {
+            $table->dropColumn('shipping_method_id');
+        });
+        Schema::table('orders', function (Blueprint $table) {
+            $table->dropColumn('customer_id');
+        });
+    }
+};